Perceptron Classifier ML model
This repo contains a Jupyter notebook and dataset for a Python implementation of the Perceptron Classifier. The algorithm is a simple linear classifier for binary classification with multiple input features. The notebook provides a guide on how to use the classifier on the included dataset with examples of training and evaluation.